# backend-ttgo/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Overvak - Sistema de monitoreo remoto para personas con Alzheimer a través de un wearable.",
    "description": "Documentación oficial de la API RESTful para el ecosistema **Overvak**. \n\nEste backend gestiona la comunicación entre los dispositivos wearables y la aplicación web. \n\n**Módulos del sistema:**\n1. **Autenticación:** Gestión de cuidadores y tokens.\n2. **Pacientes:** Vinculación de dispositivos y perfiles.\n3. **Datos IoT:** Ingesta de sensores, tales como el MAX30102, GPS (NEO6MV2) y acelerómetro.\n4. **Geocercas:** Algoritmos de punto en polígono para zonas seguras.\n5. **Alertas:** Notificaciones de emergencia.",
    "version": "1.2.1"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Introduce tu token así: Bearer <tu_token>"
    }
  },
  "paths": {
...
    "/api/geocercas/": {
      "post": {
        "tags": [
          "Geocercas"
        ],
        "description": "Crea una nueva zona segura (polígono) para un paciente.",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "pacienteId": {
                  "type": "string",
                  "example": "645..."
                },
                "nombre": {
                  "type": "string",
                  "example": "Casa"
                },
                "coords": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "example": [
                      19.4325,
                      -99.133
                    ],
                    "items": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Geocerca creada"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Datos inválidos (mínimo 3 puntos)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geocercas/{pacienteId}": {
      "get": {
        "tags": [
          "Geocercas"
        ],
        "description": "Obtiene todas las zonas seguras de un paciente.",
        "parameters": [
          {
            "name": "pacienteId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID del paciente"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geocercas/{geocercaId}": {
      "put": {
        "tags": [
          "Geocercas"
        ],
        "description": "Actualiza el nombre o las coordenadas de una geocerca.",
        "parameters": [
          {
            "name": "geocercaId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID de la geocerca"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "type": "string",
                  "example": "Casa Nueva"
                },
                "coords": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "example": [
                      19.3,
                      -99.3
                    ],
                    "items": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Geocercas"
        ],
        "description": "Elimina una zona segura.",
        "parameters": [
          {
            "name": "geocercaId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID de la geocerca"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/geocercas/verificar": {
      "post": {
        "tags": [
          "Geocercas"
        ],
        "description": "Endpoint lógico para el ESP32. Verifica si la coordenada actual está dentro de alguna geocerca y genera alertas.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "dispositivo_id": {
                  "type": "string",
                  "example": "a1b2c3d4..."
                },
                "lat": {
                  "type": "string",
                  "example": "19.4326 N"
                },
                "lng": {
                  "type": "string",
                  "example": "99.1322 W"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verificación completada",
            "schema": {
              "type": "object",
              "properties": {
                "inside": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}